add ability to ignore migrations
