+++ src/FavoriteServiceProvider.php
@@ -8,23 +8,31 @@ class FavoriteServiceProvider extends ServiceProvider
 {
     public function boot()
     {
-        $this->publishes([
-            \dirname(__DIR__).'/config/favorite.php' => config_path('favorite.php'),
-        ], 'favorite-config');
+        if (app()->runningInConsole()) {
+            $this->registerMigrations();
 
-        $this->publishes([
-            \dirname(__DIR__).'/migrations/' => database_path('migrations'),
-        ], 'favorite-migrations');
+            $this->publishes([
+                __DIR__ . '/../migrations/' => database_path('migrations'),
+            ], 'favorite-migrations');
 
-        if ($this->app->runningInConsole()) {
-            $this->loadMigrationsFrom(\dirname(__DIR__).'/migrations/');
+            $this->publishes([
+                __DIR__ . '/../config/favorite.php' => config_path('favorite.php'),
+            ], 'favorite-config');
+
+        }
+    }
+
+    protected function registerMigrations()
+    {
+        if (Favorite::shouldRunMigrations()) {
+            return $this->loadMigrationsFrom(\dirname(__DIR__) . '/migrations/');
         }
     }
 
     public function register()
     {
         $this->mergeConfigFrom(
-            \dirname(__DIR__).'/config/favorite.php',
+            \dirname(__DIR__) . '/config/favorite.php',
             'favorite'
         );
     }
